refactor: reorganiza estrutura e corrige imports
