+++ components/onboarding/OnboardingScreen.tsx
@@ -1,4 +1,4 @@
-import data, { OnboardingData } from "@/data/data";
+import data, { OnboardingData } from "@/src/data/data";
 import React from "react";
 import { FlatList, StyleSheet, View, ViewToken } from "react-native";
 import Animated, {
